+++ CHANGELOG.md
@@ -23,6 +23,7 @@ v0.15.19 (??.??.????)
  * new Wrapper function tx_rnbase_util_Logger::devLog
  * new Wrapper function tx_rnbase_mod_Util::mapRecordTypeToSpriteIconName
  * new Wrapper function tx_rnbase_parameters::setGetParameter
+ * new Wrapper function tx_rnbase_mod_BaseModule::issueCommand
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall